Extract DynamicServiceLogic into object
- Rework port assignment logic to use iterators
- Fix bug in which config.localPortMax was not included (Range.apply
  gives exclusive range)
